requirements.txt
python-can
keyboard